+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod derivation;
 pub mod error;
 pub mod event;
 pub mod event_message;
+pub mod keri;
 pub mod log;
 pub mod prefix;
 pub mod signer;
